requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
distlib==0.3.9
Django==5.1.6
django-ckeditor==6.7.2
django-ckeditor-5==0.2.17
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-import-export==4.3.7
django-js-asset==3.1.2
django-mathfilters==1.0.0
django-taggit==6.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Faker==37.1.0
filelock==3.17.0
gunicorn==23.0.0
mysql-connector-python==9.1.0
numpy==2.2.2
packaging==25.0
pd==0.0.4
pillow==11.0.0
platformdirs==4.3.6
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.1.1
shortuuid==1.0.13
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.1
virtualenv==20.29.2

# ✅ Add these 2 lines for Cloudinary support:
cloudinary==1.40.0
django-cloudinary-storage==0.3.0
